Chore : Move some types around:
